+++ editor/src/canvas/Clipboard.cpp
@@ -11,8 +11,8 @@ namespace pi {
 
 Clipboard::Clipboard(sf::Vector2<uint32_t> dimensions) : dimensions(dimensions) {}
 
-void Clipboard::write_to_clipboard(uint8_t value, size_t i, size_t j, size_t layer) {
-	while (layer >= cell_values.size()) { cell_values.push_back(std::vector<uint8_t>{}); }
+void Clipboard::write_to_clipboard(uint32_t value, size_t i, size_t j, size_t layer) {
+	while (layer >= cell_values.size()) { cell_values.push_back(std::vector<uint32_t>{}); }
 	for (auto u{0}; u < dimensions.x; ++u) {
 		for (auto v{0}; v < dimensions.y; ++v) { cell_values.at(layer).push_back(0); }
 	}
@@ -23,8 +23,8 @@ void Clipboard::write_to_clipboard(uint8_t value, size_t i, size_t j, size_t lay
 
 void Clipboard::clear_clipboard() { cell_values.clear(); }
 
-uint8_t Clipboard::get_value_at(size_t i, size_t j) { return cell_values.back().at(i + j * dimensions.x); }
+uint32_t Clipboard::get_value_at(size_t i, size_t j) { return cell_values.back().at(i + j * dimensions.x); }
 
-uint8_t Clipboard::get_value_at(size_t i, size_t j, size_t layer) { return cell_values.at(layer).at(i + j * dimensions.x); }
+uint32_t Clipboard::get_value_at(size_t i, size_t j, size_t layer) { return cell_values.at(layer).at(i + j * dimensions.x); }
 
 }
